Consolidate player modules
