+++ pages/players/index.tsx
@@ -1,22 +1,7 @@
-import { Box, Typography } from "@mui/material";
-import React from "react";
+import type { NextPage } from "next";
 
-import PlayerSearch from "../../components/players/Search";
+import { Browse } from "modules/players";
 
-const Players = () => {
-  return (
-    <Box>
-      <Typography variant="body1">
-        Players are the basic building blocks of you board game statistics.
-        Users have a &quot;main player&quot; created for them as soon as they
-        register with our app. This player is used to track <i>your</i>{" "}
-        statistics. However, you can create other players for people that
-        don&apos;t want to have an account at OBGS. These are called supervised
-        players and all the management for them is done here.
-      </Typography>
-      <PlayerSearch />
-    </Box>
-  );
-};
+const BrowsePlayers: NextPage = () => <Browse />;
 
-export default Players;
+export default BrowsePlayers;
